Added scripts for running uclust, parsing output
Added a shell script (mlti_strngth_uclust.sh) to run uclust with a
variety of identity thresholds.  Also, a script (get_clust.*) to parse
the cluster members into a Lek-like-output in both Awk and Python.  I
couldn't get the Awk script to work, but the Python works
fantastically, by the looks of it.